refactor : CommunityMain 패키지 정리
