+++ feature/community/src/main/kotlin/com/wepli/community/main/screen/CommunityScreen.kt
@@ -1,4 +1,4 @@
-package com.wepli.community.main
+package com.wepli.community.main.screen
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.Image
@@ -23,6 +23,7 @@ import appbar.AppBarIconType
 import appbar.WepliAppBar
 import com.wepli.community.component.PostItem
 import com.wepli.community.component.WePLiStoryLayout
+import com.wepli.community.main.viewmodel.CommunityViewModel
 import com.wepli.designsystem.R
 import com.wepli.shared.feature.uimodel.user.UserUiData
 import com.wepli.shared.feature.uimodel.community.PostUiData
